chore: deflake e2e_reqresp tests
The issue was racing on network readiness/state sync and occasionally sending to the wrong proposer (attester ordering mismatch).

  Fixes:

  - Wait for full P2P mesh connectivity instead of a fixed sleep.
  - Wait for all validator nodes to catch up to the block where the funded account was deployed before preparing/sending txs.
  - Align proposer selection to the validator order used to create nodes.
  - Await sendTx and fail fast on errors to avoid silent drops

NOTE:
The plan is to delete the `no_handhsake` version, but in the meantime, there is significant code duplication between the 2 (they are essentially identical), and any change to one test would have to be copied over to the other.
I've decided to clean up this a bit to make our life easier.
